* theme.c (Theme_Build_Objects): Use style 'gpe_labels', not 'gpe_ownerinfo_labels'
* ipaqscreen/main.c (ipaqscreen_Build_Objects): Make the GUI a bit
nicer.  Wrap a scrolledwindow/viewport around table and return
that.  Fix padding and justification. Use gpe_labels style for the
labels in the left table column.

* theme.c (Theme_Build_Objects): Make the GUI a bit nicer.  Wrap a
scrolledwindow/viewport around table and return that.  Fix padding
and justification. Use gpe_ownerinfo_labels style for the labels
in the left table column.


git-svn-id: svn://projects.linuxtogo.org/svn/gpe/trunk@1556 d4735bf7-8e1f-0410-bce4-ef584eb11c25